requirements.txt
flask==3.0.2
flask-cors==4.0.0
pillow==10.2.0
numpy==1.26.4
torch==2.2.0
torchvision==0.17.0
opencv-python==4.9.0.80
python-dotenv==1.0.1
gunicorn==21.2.0 